% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Map Fatality Analysis Reporting System data}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{state.num}{A character or integer value of the state number for which
the accidents shall be mapped.}

\item{years}{A character or integer value of the year for which the data shall
be read and mapped.}
}
\value{
This function returns a map with the geo-locations of the accidents
that happened in the requested state and year.
}
\description{
This function creates a map with the geo-locations of the accidents for the
given state and given year.
}
\details{
The function calls the functions \code{\link{make_filename}} and
\code{\link{fars_read}}.
}
\note{
The function execution is stopped if an invalid state number or year is
given or when there is no data to plot.
}
\examples{
fars_map_state(1, 2015)
fars_map_state("1", "2015")

}
