% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read}
\alias{fars_read}
\title{Read a Fatality Analysis Reporting System data file}
\usage{
fars_read(filename)
}
\arguments{
\item{filename}{A character string giving the name of the data file.}
}
\value{
This function returns the read data in a \code{tibble} format.
}
\description{
This function reads a Fatality Analysis Reporting System .csv data file and
returns the data in a \code{tibble} format.
}
\note{
The function execution is stopped if no file with the provided
filename exists in the working directory.
}
\examples{
\dontrun{
        accident_2013 <- fars_read("accident_2013.csv.bz2")
        }

}
