% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Read multiple Fatality Analysis Reporting System "csv.bz2"-files}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{A character or integer vector that contains the years from which
the data shall be read.}
}
\value{
This function returns a list of tibbles that contain the month and
year of the original data entries.
}
\description{
This function reads multiple Fatality Analysis Reporting System "csv.bz2"-files
and returns a list with the corresponding tibbles that contain the month and
year of the original data entries. The parameter \code{years} specifies from
which years the data shall be read.
}
\details{
The function calls the functions \code{\link{make_filename}} and
\code{\link{fars_read}}.
}
\note{
The function execution is stopped if the filename creation fails or if
no corresponding file for the generated filename exists in the working
directory.
}
\examples{
fars_month_year <- fars_read_years(c(2013, 2014, 2015))
fars_month_year <- fars_read_years(c("2013", "2014", "2015"))

}
