% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Summarize Fatality Analysis Reporting System data}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{A character or integer vector that contains the years from which
the data shall be read and summarized.}
}
\value{
This function returns a tibble in the wide-format with one column per
year and one row per month. The values are the corresponding counts of
incidents.
}
\description{
This function summarizes multiple Fatality Analysis Reporting System
"csv.bz2"-files after reading them in. For every year the count of incidents
per month is returned. The parameter \code{years} specifies from which years
the data shall be read.
}
\details{
The function calls the function \code{\link{fars_read_years}}.
}
\examples{
\dontrun{
        fars_summarize_years(c(2013, 2014, 2015))
        fars_summarize_years(c("2013", "2014", "2015"))
        }

}
