% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_functions.R
\name{make_filename}
\alias{make_filename}
\title{Make a filename for a Fatality Analysis Reporting System "csv.bz2"-file}
\usage{
make_filename(year)
}
\arguments{
\item{year}{A character or integer vector with the years that are
displayed in the filenames.}
}
\value{
This function returns a character vector with the created filenames.
}
\description{
This function creates a character vector to create a Fatality Analysis
Reporting System filename. The single values consists of "accident", the
parameter \code{year} and the ending "csv.bz2".
}
\note{
The function execution is stopped if the given parameter cannot be
coerced into a integer value or vector.
}
\examples{
filename_2013 <- make_filename("2013")
filename_2013_2014 <- make_filename(c(2013, 2014))

}
